refactor: 네이밍 변경
motivation:
- 패키지 네이밍 컨벤션, 클래스 이름의 의미 중복 및 컨벤션을 따르기 위해 수정했습니다.

modification:
- 패키지명을 snake_case -> camelCase 로 변경
- RestController 를 사용하는 컨트롤러는 이름도 Rest를 붙이게 변경
- 클래스 네이밍에서 의미가 중복되는 단어를 제거
