+++ src/main/java/com/aroom/domain/reservation/service/ReservationService.java
@@ -1,13 +1,13 @@
 package com.aroom.domain.reservation.service;
 
 import com.aroom.domain.reservation.dto.request.ReservationRequest;
-import com.aroom.domain.reservation.dto.response.ReservedReservationResponse;
+import com.aroom.domain.reservation.dto.response.ReservationResponse;
 import org.springframework.stereotype.Service;
 
 @Service
 public class ReservationService {
 
-    public ReservedReservationResponse reserveRoom(ReservationRequest request) {
+    public ReservationResponse reserveRoom(ReservationRequest request) {
         return null;
     }
 }
