+++ src/main/java/com/aroom/domain/reservationRoom/repository/ReservationRoomRepository.java
@@ -1,6 +1,6 @@
-package com.aroom.domain.reservation_room.repository;
+package com.aroom.domain.reservationRoom.repository;
 
-import com.aroom.domain.reservation_room.model.ReservationRoom;
+import com.aroom.domain.reservationRoom.model.ReservationRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReservationRoomRepository extends JpaRepository<ReservationRoom, Long> {
